+++ HISTORY.md
@@ -10,6 +10,26 @@
 ********************************************************************************
 ## NEWS
 
+0.2.14(2025.03.03 16.16.50)
+------------------------------
+- [ini] add ConfigParserMod  
+- [AttrKit] add  
+- [Text]:  
+	- add parse__dict_ini  
+	- fix__json/parse__json  
+	- add delete__cmts_c  
+	- add search__group - need fix  
+- [TextFile] add DictTextFileLoader  
+- [NestInit_AttrsLambdasResolve] use with Resolvers  
+- [Nest*_Attrs]:  
+	- separate all for Attrs Contains/Eq/Len/StR  
+	- [NestSAI_AttrAnycase] deprecate all Setattrs* cause recursion exx  
+- [EqValid]:  
+	- add EqValid_AnnotsAllExists  
+	- add EqValid_AttrsByObjNotPrivate*  
+- [PV] deprecate old and add new PvLoaderIni/json/env  
+- [Kwargs] ref to simple dict nesting  
+
 0.2.13(2025.02.27 15.07.31)
 ------------------------------
 - [serial] turn back ValueUnit